CRUD Room Modal

+++ booking_frontend/src/App.vue
@@ -2,18 +2,24 @@
 import RoomItem from './components/RoomItem.vue'
 import CreateRoomModal from './components/CreateRoomModal.vue'
 import FacItem from './components/FacItem.vue'
+import DetailRoomModal from './components/DetailRoomModal.vue'
+import UpdateRoomModal from './components/UpdateRoomModal.vue'
+import * as yup from "yup";
 
 export default {
   name: 'App',
   components: {
-    RoomItem, CreateRoomModal, FacItem
+    RoomItem, CreateRoomModal, FacItem, DetailRoomModal, UpdateRoomModal
   },
   data() {
     return {
       title: 'Hotel Booking',
       listRoom: [],
-      formisOpen: false,
-      listFacs: []
+      isModalCreateRoomOpen: false,
+      listFacs: [],
+      errors: {
+        title: 'Is required'
+      }
     }
   },
   methods: {
@@ -22,32 +28,68 @@ export default {
       const finalRes = await res.json();
       this.listRoom = finalRes;
     },
+
     async createRoom(data) {
+      const yupObject = yup.object().shape({
+        name: yup.string().required(),
+        description: yup.string(),
+        price: yup.number(),
+        roomsize: yup.number(),
+        facilities: yup.array()
+          .of(
+            yup.object().shape({
+              name: yup.string(),
+              checked: yup.boolean(),
+            })
+          )
+          .compact((v) => !v.checked),
+        image: yup.string()
+      });
+      yupObject.validate(data);
       try {
-        const res = await fetch("http://127.0.0.1:8000/rooms/", {
+        // TODO:
+        // validate data use yup validator
+        // if correct -> send request + push into listRoom
+        // incorrect -> save error message into state + show error message under corresponding input
+        await fetch("http://127.0.0.1:8000/rooms/", {
           method: "POST",
           headers: {
             "Content-Type": "application/json",
           },
           body: JSON.stringify(data)
         })
+        data.facilities = this.listFacs.filter(e => data.facilities.includes(e.name))
         this.listRoom.push(data)
-        console.log('====', this.listRoom)
-        // TODO: apppend res to this.listRoom
+        this.toggleModalCreateRoom()
       } catch (e) {
-        console.error(e)
-        // custom logic handle error
+        alert(e.message)
       }
     },
+
+    async updateRoom(id, data) {
+      try {
+        fetch(`http://127.0.0.1:8000/rooms/${id}`, {
+          method: 'PUT',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify(data)
+        })
+      }
+      catch (e) {
+        alert(e.message)
+      }
+    },
+
     deleteRoom(id) {
-      alert('aze')
+      if (!confirm('are u sure?')) return
       fetch(`http://127.0.0.1:8000/rooms/${id}`, {
         method: 'DELETE',
       })
       this.listRoom = this.listRoom.filter(room => room.id !== id)
     },
-    toggleForm() {
-      this.formisOpen = !this.formisOpen;
+    toggleModalCreateRoom() {
+      this.isModalCreateRoomOpen = !this.isModalCreateRoomOpen;
     },
     async getFacs() {
       const res = await fetch("http://127.0.0.1:8000/facility/");
@@ -65,22 +107,50 @@ export default {
 <template>
   <h1>{{ title }}</h1>
   <div v-for="room in listRoom">
-    <RoomItem :room="room" :deleteRoom="deleteRoom"/>
+    <RoomItem :room="room" :deleteRoom="deleteRoom" :updateRoom="updateRoom" :listFacs="listFacs" />
   </div>
-  <div v-if="formisOpen">
-    <CreateRoomModal :createRoom="createRoom" :formisOpen="formisOpen" :listFacs="listFacs" @click.self="toggleForm" />
+  <div v-if="isModalCreateRoomOpen">
+    <CreateRoomModal :createRoom="createRoom" :isOpen="isModalCreateRoomOpen" :listFacs="listFacs"
+      @click.self="toggleModalCreateRoom" />
   </div>
-  <button @click="toggleForm">Create new room</button>
+  <!-- <div>
+    <UpdateRoomModal :updateRoom="updateRoom" />
+  </div> -->
+  <button class="create-button" @click="toggleModalCreateRoom">Create new room</button>
 </template>
 
 
 <style scope>
 body {
   padding: 0;
-  border: 1px solid purple
 }
 
 h1 {
-  color: pink;
+  background: linear-gradient(90deg, #ff0000, #ffff00, #ff00f3, #0033ff, #ff00c4, #ff0000);
+  background-size: 400%;
+  font-size: 70px;
+  font-family: sans-serif;
+  letter-spacing: 5px;
+  font-weight: 600;
+  word-spacing: 5px;
+  -webkit-text-fill-color: transparent;
+  -webkit-background-clip: text;
+  animation: animate 20s linear infinite;
+}
+
+@keyframes animate {
+  0% {
+    background-position: 0%;
+  }
+
+  10% {
+    background-position: 400%;
+  }
+}
+
+.create-button {
+  margin-top: 1%;
+  background-color: blueviolet;
+  border: 1px solid purple;
 }
 </style>
